chatbot_rag-main/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
python-dotenv==1.1.1
pydantic==2.11.7
langchain==0.3.27
langchain-community==0.3.27
chromadb==1.0.15
sentence-transformers==5.0.0
llama-cpp-python==0.3.14
pandas==2.3.1
tqdm==4.67.1
tabulate==0.9.0
rich==14.1.0
google-generativeai==0.8.5 